Use new `MachineAutostop` type with option for autosuspend
After discussion, we decided that the best way to configure autosuspend
is through the existing Machine autostop field. Consequently, this
introduces a textual representation for the field with three settings:
"off", "stop", and "suspend".

To maintain backward compatibility for existing Machines API clients,
JSON deserialization will continue to accept booleans, and JSON
serialization will prefer the boolean representations for "off" and
"stop".

When this change is incorporated into flyctl, `fly.toml` serialization
will default to using the new textual representation. Deserialization
will continue to accept booleans. Through a quirk of the parser, it will
also accept the strings "false" and "true"; unfortunately, I don't think
we'll be able to avoid that without using go-toml's unstable API.